requirements.txt
langchain
langchain-community
openai
tiktoken
utils
faiss-cpu
langchain-openai
Flask
python-dotenv